pyproject.toml
[project]
name = "what-have-i-done"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.99.6",
    "python-dotenv>=1.1.1",
]
